+++ include/CausalDilatedConv1d.h
@@ -9,10 +9,10 @@ namespace MicroTorch
     class CausalDilatedConv1d
     {
     public:
-        CausalDilatedConv1d(int in_channels, int out_channels, int kernel_size, bool bias, int dilation) : m_inChannels(in_channels), m_outChannels(out_channels), 
+        CausalDilatedConv1d(size_t in_channels, size_t out_channels, size_t kernel_size, bool bias, size_t dilation) : m_inChannels(in_channels), m_outChannels(out_channels), 
             m_kernelSize(kernel_size), m_bias(bias), m_dilation(dilation), m_internalPadding((dilation * (kernel_size - 1)) / 2), m_b(Eigen::RowVectorXf::Zero(out_channels))
         {
-            for(int i = 0; i < out_channels; i++)
+            for(size_t i = 0; i < out_channels; i++)
             {
                 m_w.push_back( RowMatrixXf::Zero(in_channels, kernel_size) );
                 m_dilatedW.push_back( RowMatrixXf::Zero(in_channels, dilation * (kernel_size - 1) + 1) );
@@ -20,14 +20,14 @@ namespace MicroTorch
         }
         ~CausalDilatedConv1d() = default;
 
-        void setWeight(int channel, const Eigen::Ref<RowMatrixXf>& m)
+        void setWeight(size_t channel, const Eigen::Ref<RowMatrixXf>& m)
         {
             assert(m.rows() == m_inChannels);
             assert(m.cols() == m_kernelSize);
             m_w[channel] = m;
 
             // Calculate dilated weights
-            for(int i = 0; i < m.rows(); i++)
+            for(size_t i = 0; i < m.rows(); i++)
             {
                 RowMatrixXf row = m.row(i);
                 m_dilatedW[channel].row(i) = dilate( row, m_dilation );
@@ -43,9 +43,9 @@ namespace MicroTorch
         inline RowMatrixXf forward( const Eigen::Ref<RowMatrixXf>& x ) const noexcept
         {    
             RowMatrixXf y = RowMatrixXf::Zero(m_outChannels, x.cols());
-            for(int i = 0; i < m_outChannels; i++)
+            for(size_t i = 0; i < m_outChannels; i++)
             {
-                for(int j = 0; j < m_inChannels; j++)
+                for(size_t j = 0; j < m_inChannels; j++)
                     if( m_internalPadding > 0)
                     {
                         RowMatrixXf row = x.row(j);
@@ -65,23 +65,23 @@ namespace MicroTorch
             return y;
         }
  
-        int getInChannels() const { return m_inChannels; }
-        int getOutChannels() const { return m_outChannels; }
-        int getKernelSize() const { return m_kernelSize; }
-        int getBias() const { return m_bias; }
-        int getDilation() const { return m_dilation; }
+        size_t getInChannels() const { return m_inChannels; }
+        size_t getOutChannels() const { return m_outChannels; }
+        size_t getKernelSize() const { return m_kernelSize; }
+        size_t getBias() const { return m_bias; }
+        size_t getDilation() const { return m_dilation; }
         
         void loadStateDict(std::map<std::string, nlohmann::json> state_dict)
         {
             auto w = loadTensor( std::string("weight"), state_dict );
             auto b = loadVector( std::string("bias"), state_dict );
-            for(int i = 0; i < m_outChannels; i++)
+            for(size_t i = 0; i < m_outChannels; i++)
                 setWeight( i, w[i] );
             setBias( b );
         }
 
     private:
-        int m_inChannels, m_outChannels, m_kernelSize, m_dilation, m_internalPadding;
+        size_t m_inChannels, m_outChannels, m_kernelSize, m_dilation, m_internalPadding;
         bool m_bias;
 
         std::vector<RowMatrixXf> m_w, m_dilatedW; // W = [Outs, Ins, Kernel]
