+++ include/Conv1d.h
@@ -9,15 +9,15 @@ namespace MicroTorch
     class Conv1d
     {
     public:
-        Conv1d(int in_channels, int out_channels, int kernel_size, bool bias) : m_inChannels(in_channels), m_outChannels(out_channels), 
+        Conv1d(size_t in_channels, size_t out_channels, size_t kernel_size, bool bias) : m_inChannels(in_channels), m_outChannels(out_channels), 
             m_kernelSize(kernel_size), m_bias(bias),m_b(Eigen::RowVectorXf::Zero(out_channels))
         {
-            for(int i = 0; i < out_channels; i++)
+            for(size_t i = 0; i < out_channels; i++)
                 m_w.push_back( RowMatrixXf::Zero(in_channels, kernel_size) );
         }
         ~Conv1d() = default;
 
-        void setWeight(int channel, const Eigen::Ref<RowMatrixXf>& m)
+        void setWeight(size_t channel, const Eigen::Ref<RowMatrixXf>& m)
         {
             assert(m.rows() == m_inChannels);
             assert(m.cols() == m_kernelSize);
@@ -33,9 +33,9 @@ namespace MicroTorch
         inline RowMatrixXf forward( const Eigen::Ref<RowMatrixXf>& x ) const noexcept
         {
             RowMatrixXf y = RowMatrixXf::Zero(m_outChannels, x.cols());
-            for(int i = 0; i < m_outChannels; i++)
+            for(size_t i = 0; i < m_outChannels; i++)
             {
-                for(int j = 0; j < m_inChannels; j++)
+                for(size_t j = 0; j < m_inChannels; j++)
                 {
                     RowMatrixXf row = x.row(j);
                     RowMatrixXf weights = m_w[i].row(j);
@@ -47,22 +47,22 @@ namespace MicroTorch
             return y;
         }
 
-        int getInChannels() const { return m_inChannels; }
-        int getOutChannels() const { return m_outChannels; }
-        int getKernelSize() const { return m_kernelSize; }
-        int getBias() const { return m_bias; }
+        size_t getInChannels() const { return m_inChannels; }
+        size_t getOutChannels() const { return m_outChannels; }
+        size_t getKernelSize() const { return m_kernelSize; }
+        size_t getBias() const { return m_bias; }
         
         void loadStateDict(std::map<std::string, nlohmann::json> state_dict)
         {
             auto w = loadTensor( std::string("weight"), state_dict );
             auto b = loadVector( std::string("bias"), state_dict );
-            for(int i = 0; i < m_outChannels; i++)
+            for(size_t i = 0; i < m_outChannels; i++)
                 setWeight( i, w[i] );
             setBias( b );
         }
 
     private:
-        int m_inChannels, m_outChannels, m_kernelSize;
+        size_t m_inChannels, m_outChannels, m_kernelSize;
         bool m_bias;
 
         std::vector<RowMatrixXf> m_w; // W = [Outs, Ins, Kernel]
