+++ include/models/ResRNN.h
@@ -13,7 +13,7 @@ namespace MicroTorch
     class ResRNN : public BaseModel
     {
     public:
-        ResRNN(int input_size, int hidden_size, int output_size, bool rnn_bias, bool linear_bias, float norm_mean, float norm_std) : BaseModel(norm_mean, norm_std), 
+        ResRNN(size_t input_size, size_t hidden_size, size_t output_size, bool rnn_bias, bool linear_bias, float norm_mean, float norm_std) : BaseModel(norm_mean, norm_std), 
             m_rnn(input_size, hidden_size, rnn_bias), 
             m_linear(hidden_size, output_size, linear_bias) {}
         ~ResRNN() = default;
